+++ xmtp_mls/src/lib.rs
@@ -1,9 +1,9 @@
 pub mod api_client_wrapper;
-pub mod association;
 pub mod builder;
 pub mod client;
 pub mod codecs;
 mod configuration;
+pub mod credential;
 pub mod groups;
 pub mod identity;
 pub mod mock_xmtp_api_client;
